--- src/main/java/Selenium2Library.java
@@ -1,19 +1,0 @@
-public class Selenium2Library extends com.github.markusbernhardt.selenium2library.Selenium2Library {
-
-	public Selenium2Library(String timeout, String implicitWait, String runOnFailure) {
-		super(timeout, implicitWait, runOnFailure);
-	}
-
-	public Selenium2Library(String timeout, String implicitWait) {
-		super(timeout, implicitWait);
-	}
-
-	public Selenium2Library(String timeout) {
-		super(timeout);
-	}
-
-	public Selenium2Library() {
-		super();
-	}
-
-}
