+++ src/main/java/SeleniumLibrary.java
@@ -1,0 +1,19 @@
+public class SeleniumLibrary extends com.github.markusbernhardt.seleniumlibrary.Selenium2Library {
+
+	public SeleniumLibrary(String timeout, String implicitWait, String runOnFailure) {
+		super(timeout, implicitWait, runOnFailure);
+	}
+
+	public SeleniumLibrary(String timeout, String implicitWait) {
+		super(timeout, implicitWait);
+	}
+
+	public SeleniumLibrary(String timeout) {
+		super(timeout);
+	}
+
+	public SeleniumLibrary() {
+		super();
+	}
+
+}
