+++ src/main/java/com/github/markusbernhardt/seleniumlibrary/keywords/BrowserManagement.java
@@ -1,4 +1,4 @@
-package com.github.markusbernhardt.selenium2library.keywords;
+package com.github.markusbernhardt.seleniumlibrary.keywords;
 
 import io.appium.java_client.ios.IOSDriver;
 import io.selendroid.client.SelendroidDriver;
@@ -46,14 +46,14 @@ import org.robotframework.javalib.annotation.RobotKeyword;
 import org.robotframework.javalib.annotation.RobotKeywordOverload;
 import org.robotframework.javalib.annotation.RobotKeywords;
 
-import com.github.markusbernhardt.selenium2library.RunOnFailureKeywordsAdapter;
-import com.github.markusbernhardt.selenium2library.Selenium2LibraryFatalException;
-import com.github.markusbernhardt.selenium2library.Selenium2LibraryNonFatalException;
-import com.github.markusbernhardt.selenium2library.locators.ElementFinder;
-import com.github.markusbernhardt.selenium2library.locators.WindowManager;
-import com.github.markusbernhardt.selenium2library.utils.Robotframework;
-import com.github.markusbernhardt.selenium2library.utils.WebDriverCache;
-import com.github.markusbernhardt.selenium2library.utils.WebDriverCache.SessionIdAliasWebDriverTuple;
+import com.github.markusbernhardt.seleniumlibrary.RunOnFailureKeywordsAdapter;
+import com.github.markusbernhardt.seleniumlibrary.Selenium2LibraryFatalException;
+import com.github.markusbernhardt.seleniumlibrary.Selenium2LibraryNonFatalException;
+import com.github.markusbernhardt.seleniumlibrary.locators.ElementFinder;
+import com.github.markusbernhardt.seleniumlibrary.locators.WindowManager;
+import com.github.markusbernhardt.seleniumlibrary.utils.Robotframework;
+import com.github.markusbernhardt.seleniumlibrary.utils.WebDriverCache;
+import com.github.markusbernhardt.seleniumlibrary.utils.WebDriverCache.SessionIdAliasWebDriverTuple;
 import com.opera.core.systems.OperaDriver;
 
 @SuppressWarnings("deprecation")
