+++ src/main/java/com/github/markusbernhardt/seleniumlibrary/keywords/Element.java
@@ -1,4 +1,4 @@
-package com.github.markusbernhardt.selenium2library.keywords;
+package com.github.markusbernhardt.seleniumlibrary.keywords;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -18,10 +18,10 @@ import org.robotframework.javalib.annotation.RobotKeyword;
 import org.robotframework.javalib.annotation.RobotKeywordOverload;
 import org.robotframework.javalib.annotation.RobotKeywords;
 
-import com.github.markusbernhardt.selenium2library.RunOnFailureKeywordsAdapter;
-import com.github.markusbernhardt.selenium2library.Selenium2LibraryNonFatalException;
-import com.github.markusbernhardt.selenium2library.locators.ElementFinder;
-import com.github.markusbernhardt.selenium2library.utils.Python;
+import com.github.markusbernhardt.seleniumlibrary.RunOnFailureKeywordsAdapter;
+import com.github.markusbernhardt.seleniumlibrary.Selenium2LibraryNonFatalException;
+import com.github.markusbernhardt.seleniumlibrary.locators.ElementFinder;
+import com.github.markusbernhardt.seleniumlibrary.utils.Python;
 
 @RobotKeywords
 public class Element extends RunOnFailureKeywordsAdapter {
