+++ src/main/java/com/github/markusbernhardt/seleniumlibrary/keywords/Waiting.java
@@ -1,4 +1,4 @@
-package com.github.markusbernhardt.selenium2library.keywords;
+package com.github.markusbernhardt.seleniumlibrary.keywords;
 
 import org.openqa.selenium.JavascriptExecutor;
 import org.robotframework.javalib.annotation.ArgumentNames;
@@ -7,9 +7,9 @@ import org.robotframework.javalib.annotation.RobotKeyword;
 import org.robotframework.javalib.annotation.RobotKeywordOverload;
 import org.robotframework.javalib.annotation.RobotKeywords;
 
-import com.github.markusbernhardt.selenium2library.RunOnFailureKeywordsAdapter;
-import com.github.markusbernhardt.selenium2library.Selenium2LibraryNonFatalException;
-import com.github.markusbernhardt.selenium2library.utils.Robotframework;
+import com.github.markusbernhardt.seleniumlibrary.RunOnFailureKeywordsAdapter;
+import com.github.markusbernhardt.seleniumlibrary.Selenium2LibraryNonFatalException;
+import com.github.markusbernhardt.seleniumlibrary.utils.Robotframework;
 
 @RobotKeywords
 public class Waiting extends RunOnFailureKeywordsAdapter {
